scripts/requirements.txt
paho-mqtt==1.6.1
sseclient==0.0.27